- Renamed the Events namespace to EventManagement - Consolidated the layer system from layer and overlay to just layers - Removed EventDispatcher.h and EventDispatcher.cpp as they are not needed anymore - Consolidated the different types of events into the EventTypes folder and created a header that includes all the event types.
