+++ example-user-project/src/ExampleLayer.cpp
@@ -29,9 +29,9 @@ void ExampleLayer::OnImGuiRender()
 }
 
 
-Ayla::Events::EEventCategory ExampleLayer::GetAcceptingEventFlags()
+Ayla::EventManagement::EEventCategory ExampleLayer::GetAcceptingEventFlags()
 {
-    return Ayla::Events::None;
+    return Ayla::EventManagement::None;
 }
