+++ source/Ayla/Events/EventTypes/EventTypes.h
@@ -1,0 +1,9 @@
+//
+// Created by Andrew Fagan on 7/24/24.
+//
+
+#pragma once
+
+#include "ApplicationEvent.h"
+#include "KeyEvent.h"
+#include "MouseEvent.h"
