+++ source/Ayla/Input/InputState.cpp
@@ -4,15 +4,15 @@
 
 #include "InputState.h"
 
-#include "Ayla/Events/ApplicationEvent.h"
-#include "Ayla/Events/KeyEvent.h"
-#include "Ayla/Events/MouseEvent.h"
+#include "Ayla/Events/EventTypes/ApplicationEvent.h"
+#include "Ayla/Events/EventTypes/KeyEvent.h"
+#include "Ayla/Events/EventTypes/MouseEvent.h"
 #include "Ayla/Input/Keycodes.h"
 
 
 namespace Ayla::Input {
 
-    using namespace Ayla::Events;
+    using namespace Ayla::EventManagement;
 
     SInputState* SInputState::m_Instance = nullptr;
